bugfix - cast int to double in a division operation

+++ common/src/main/java/utils/FinanceUtils.java
@@ -67,7 +67,7 @@ public class FinanceUtils {
         // startDate and endDate required to calculate appropriate risk free rate for different time periods other than yearly.
         int totalNumOfDays = Days.daysBetween(startDate.toLocalDate(), endDate.toLocalDate()).getDays();
         List<Double> returnPercents = calcAllReturns(returns);
-        double averageReturnPeriod = totalNumOfDays / returnPercents.size();
+        double averageReturnPeriod = (double)totalNumOfDays / returnPercents.size();
         double adjustedRiskFreeRate = Math.pow(1.05, (averageReturnPeriod / 365)) - 1;
 
         //Calculate necessary statistics
